+++ packages/riverpod_generator/test/async_test.dart
@@ -1,5 +1,7 @@
 // ignore_for_file: omit_local_variable_types, unused_local_variable // Just checking
 
+import 'dart:async';
+
 import 'package:riverpod/misc.dart' show ProviderBase;
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:test/test.dart';
@@ -12,7 +14,7 @@ void main() {
     () {
       final container = ProviderContainer.test();
 
-      const ProviderBase<AsyncValue<String>> provider = publicProvider;
+      final ProviderBase<AsyncValue<String>> provider = publicProvider;
       final AsyncValue<String> result = container.read(publicProvider);
 
       expect(result, const AsyncData('Hello world'));
@@ -66,7 +68,7 @@ void main() {
     () async {
       final container = ProviderContainer.test();
 
-      const FamilyFamily family = familyProvider;
+      final FamilyFamily family = familyProvider;
 
       expect(familyProvider(42, third: .42).from, familyProvider);
 
@@ -179,5 +181,12 @@ void main() {
       container.read(familyProvider(42, third: .2).future),
       completion('test'),
     );
+  });
+
+  test('Regression for Future<void> providers', () async {
+    // https://github.com/rrousselGit/riverpod/issues/4282
+    final container = ProviderContainer.test();
+
+    expect(container.read(voidClassProvider), const AsyncLoading<void>());
   });
 }