+++ packages/riverpod_generator/test/integration/async.g.dart
@@ -10,7 +10,7 @@ part of 'async.dart';
 // ignore_for_file: type=lint, type=warning
 
 @ProviderFor(generic)
-const genericProvider = GenericFamily._();
+final genericProvider = GenericFamily._();
 
 final class GenericProvider<ObjT extends num>
     extends
@@ -20,7 +20,7 @@ final class GenericProvider<ObjT extends num>
           FutureOr<List<ObjT>>
         >
     with $FutureModifier<List<ObjT>>, $FutureProvider<List<ObjT>> {
-  const GenericProvider._({required GenericFamily super.from})
+  GenericProvider._({required GenericFamily super.from})
     : super(
         argument: null,
         retry: null,
@@ -70,7 +70,7 @@ final class GenericProvider<ObjT extends num>
 String _$genericHash() => r'f530516182a22020b12fb9c3868245ad003fe8bc';
 
 final class GenericFamily extends $Family {
-  const GenericFamily._()
+  GenericFamily._()
     : super(
         retry: null,
         name: r'genericProvider',
@@ -101,11 +101,11 @@ final class GenericFamily extends $Family {
 }
 
 @ProviderFor(GenericClass)
-const genericClassProvider = GenericClassFamily._();
+final genericClassProvider = GenericClassFamily._();
 
 final class GenericClassProvider<ObjT extends num>
     extends $AsyncNotifierProvider<GenericClass<ObjT>, List<ObjT>> {
-  const GenericClassProvider._({required GenericClassFamily super.from})
+  GenericClassProvider._({required GenericClassFamily super.from})
     : super(
         argument: null,
         retry: null,
@@ -149,7 +149,7 @@ final class GenericClassProvider<ObjT extends num>
 String _$genericClassHash() => r'1ee74c45e7f5efac4db7e959f75fcf29b033c1d9';
 
 final class GenericClassFamily extends $Family {
-  const GenericClassFamily._()
+  GenericClassFamily._()
     : super(
         retry: null,
         name: r'genericClassProvider',
@@ -220,11 +220,11 @@ abstract class _$GenericClass<ObjT extends num>
 }
 
 @ProviderFor(GenericArg)
-const genericArgProvider = GenericArgFamily._();
+final genericArgProvider = GenericArgFamily._();
 
 final class GenericArgProvider<ObjT extends num>
     extends $AsyncNotifierProvider<GenericArg<ObjT>, String> {
-  const GenericArgProvider._({
+  GenericArgProvider._({
     required GenericArgFamily super.from,
     required ObjT super.argument,
   }) : super(
@@ -269,7 +269,7 @@ final class GenericArgProvider<ObjT extends num>
 String _$genericArgHash() => r'cd3043cae4ddbb54b11b76599098b07301b7a33a';
 
 final class GenericArgFamily extends $Family {
-  const GenericArgFamily._()
+  GenericArgFamily._()
     : super(
         retry: null,
         name: r'genericArgProvider',
@@ -341,12 +341,12 @@ abstract class _$GenericArg<ObjT extends num> extends $AsyncNotifier<String> {
 }
 
 @ProviderFor(public)
-const publicProvider = PublicProvider._();
+final publicProvider = PublicProvider._();
 
 final class PublicProvider
     extends $FunctionalProvider<AsyncValue<String>, String, FutureOr<String>>
     with $FutureModifier<String>, $FutureProvider<String> {
-  const PublicProvider._()
+  PublicProvider._()
     : super(
         from: null,
         argument: null,
@@ -374,12 +374,12 @@ final class PublicProvider
 String _$publicHash() => r'19bceccf795e4c3a26ad1e613fd6f41aad949e2b';
 
 @ProviderFor(_private)
-const _privateProvider = _PrivateProvider._();
+final _privateProvider = _PrivateProvider._();
 
 final class _PrivateProvider
     extends $FunctionalProvider<AsyncValue<String>, String, FutureOr<String>>
     with $FutureModifier<String>, $FutureProvider<String> {
-  const _PrivateProvider._()
+  _PrivateProvider._()
     : super(
         from: null,
         argument: null,
@@ -407,12 +407,12 @@ final class _PrivateProvider
 String _$_privateHash() => r'7f0d1ff55a21e520b8471bbabc4649b5336221d4';
 
 @ProviderFor(familyOr)
-const familyOrProvider = FamilyOrFamily._();
+final familyOrProvider = FamilyOrFamily._();
 
 final class FamilyOrProvider
     extends $FunctionalProvider<AsyncValue<String>, String, FutureOr<String>>
     with $FutureModifier<String>, $FutureProvider<String> {
-  const FamilyOrProvider._({
+  FamilyOrProvider._({
     required FamilyOrFamily super.from,
     required int super.argument,
   }) : super(
@@ -459,7 +459,7 @@ String _$familyOrHash() => r'97cce80a626e228202fa30b87c07ae8319b48023';
 
 final class FamilyOrFamily extends $Family
     with $FunctionalFamilyOverride<FutureOr<String>, int> {
-  const FamilyOrFamily._()
+  FamilyOrFamily._()
     : super(
         retry: null,
         name: r'familyOrProvider',
@@ -476,12 +476,12 @@ final class FamilyOrFamily extends $Family
 }
 
 @ProviderFor(family)
-const familyProvider = FamilyFamily._();
+final familyProvider = FamilyFamily._();
 
 final class FamilyProvider
     extends $FunctionalProvider<AsyncValue<String>, String, FutureOr<String>>
     with $FutureModifier<String>, $FutureProvider<String> {
-  const FamilyProvider._({
+  FamilyProvider._({
     required FamilyFamily super.from,
     required (
       int, {
@@ -560,7 +560,7 @@ final class FamilyFamily extends $Family
             List<String>? fifth,
           })
         > {
-  const FamilyFamily._()
+  FamilyFamily._()
     : super(
         retry: null,
         name: r'familyProvider',
@@ -591,11 +591,11 @@ final class FamilyFamily extends $Family
 }
 
 @ProviderFor(PublicClass)
-const publicClassProvider = PublicClassProvider._();
+final publicClassProvider = PublicClassProvider._();
 
 final class PublicClassProvider
     extends $AsyncNotifierProvider<PublicClass, String> {
-  const PublicClassProvider._()
+  PublicClassProvider._()
     : super(
         from: null,
         argument: null,
@@ -636,11 +636,11 @@ abstract class _$PublicClass extends $AsyncNotifier<String> {
 }
 
 @ProviderFor(_PrivateClass)
-const _privateClassProvider = _PrivateClassProvider._();
+final _privateClassProvider = _PrivateClassProvider._();
 
 final class _PrivateClassProvider
     extends $AsyncNotifierProvider<_PrivateClass, String> {
-  const _PrivateClassProvider._()
+  _PrivateClassProvider._()
     : super(
         from: null,
         argument: null,
@@ -681,11 +681,11 @@ abstract class _$PrivateClass extends $AsyncNotifier<String> {
 }
 
 @ProviderFor(FamilyOrClass)
-const familyOrClassProvider = FamilyOrClassFamily._();
+final familyOrClassProvider = FamilyOrClassFamily._();
 
 final class FamilyOrClassProvider
     extends $AsyncNotifierProvider<FamilyOrClass, String> {
-  const FamilyOrClassProvider._({
+  FamilyOrClassProvider._({
     required FamilyOrClassFamily super.from,
     required int super.argument,
   }) : super(
@@ -732,7 +732,7 @@ final class FamilyOrClassFamily extends $Family
           FutureOr<String>,
           int
         > {
-  const FamilyOrClassFamily._()
+  FamilyOrClassFamily._()
     : super(
         retry: null,
         name: r'familyOrClassProvider',
@@ -771,11 +771,11 @@ abstract class _$FamilyOrClass extends $AsyncNotifier<String> {
 }
 
 @ProviderFor(FamilyClass)
-const familyClassProvider = FamilyClassFamily._();
+final familyClassProvider = FamilyClassFamily._();
 
 final class FamilyClassProvider
     extends $AsyncNotifierProvider<FamilyClass, String> {
-  const FamilyClassProvider._({
+  FamilyClassProvider._({
     required FamilyClassFamily super.from,
     required (
       int, {
@@ -835,7 +835,7 @@ final class FamilyClassFamily extends $Family
             List<String>? fifth,
           })
         > {
-  const FamilyClassFamily._()
+  FamilyClassFamily._()
     : super(
         retry: null,
         name: r'familyClassProvider',
@@ -912,11 +912,11 @@ abstract class _$FamilyClass extends $AsyncNotifier<String> {
 }
 
 @ProviderFor(Regression3490)
-const regression3490Provider = Regression3490Family._();
+final regression3490Provider = Regression3490Family._();
 
 final class Regression3490Provider<ModelT, SortT, CursorT>
     extends $NotifierProvider<Regression3490<ModelT, SortT, CursorT>, void> {
-  const Regression3490Provider._({
+  Regression3490Provider._({
     required Regression3490Family super.from,
     required ({
       String type,
@@ -975,7 +975,7 @@ final class Regression3490Provider<ModelT, SortT, CursorT>
 String _$regression3490Hash() => r'1e22e082b06069c176814aea015c20723959b1f8';
 
 final class Regression3490Family extends $Family {
-  const Regression3490Family._()
+  Regression3490Family._()
     : super(
         retry: null,
         name: r'regression3490Provider',
@@ -1070,5 +1070,49 @@ abstract class _$Regression3490<ModelT, SortT, CursorT>
               Object?
             >;
     element.handleValue(ref, null);
+  }
+}
+
+@ProviderFor(VoidClass)
+final voidClassProvider = VoidClassProvider._();
+
+final class VoidClassProvider extends $AsyncNotifierProvider<VoidClass, void> {
+  VoidClassProvider._()
+    : super(
+        from: null,
+        argument: null,
+        retry: null,
+        name: r'voidClassProvider',
+        isAutoDispose: true,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
+
+  @override
+  String debugGetCreateSourceHash() => _$voidClassHash();
+
+  @$internal
+  @override
+  VoidClass create() => VoidClass();
+}
+
+String _$voidClassHash() => r'3c3109f3474209a42efd97c7315171f74669aa49';
+
+abstract class _$VoidClass extends $AsyncNotifier<void> {
+  FutureOr<void> build();
+  @$mustCallSuper
+  @override
+  void runBuild() {
+    final created = build();
+    final ref = this.ref as $Ref<AsyncValue<void>, void>;
+    final element =
+        ref.element
+            as $ClassProviderElement<
+              AnyNotifier<AsyncValue<void>, void>,
+              AsyncValue<void>,
+              Object?,
+              Object?
+            >;
+    element.handleValue(ref, created);
   }
 }