+++ packages/riverpod_lint_flutter_test/test/lints/protected_notifier_properties.g.dart
@@ -10,10 +10,10 @@ part of 'protected_notifier_properties.dart';
 // ignore_for_file: type=lint, type=warning
 
 @ProviderFor(A)
-const aProvider = AProvider._();
+final aProvider = AProvider._();
 
 final class AProvider extends $NotifierProvider<A, int> {
-  const AProvider._()
+  AProvider._()
     : super(
         from: null,
         argument: null,
@@ -62,10 +62,10 @@ abstract class _$A extends $Notifier<int> {
 }
 
 @ProviderFor(A2)
-const a2Provider = A2Provider._();
+final a2Provider = A2Provider._();
 
 final class A2Provider extends $NotifierProvider<A2, int> {
-  const A2Provider._()
+  A2Provider._()
     : super(
         from: null,
         argument: null,
@@ -114,19 +114,17 @@ abstract class _$A2 extends $Notifier<int> {
 }
 
 @ProviderFor(A3)
-const a3Provider = A3Family._();
+final a3Provider = A3Family._();
 
 final class A3Provider extends $NotifierProvider<A3, int> {
-  const A3Provider._({
-    required A3Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a3Provider',
-         isAutoDispose: true,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A3Provider._({required A3Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a3Provider',
+        isAutoDispose: true,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a3Hash();
@@ -165,7 +163,7 @@ String _$a3Hash() => r'2e21e9af8b67b5412611e0d23b862ead56deb8e1';
 
 final class A3Family extends $Family
     with $ClassFamilyOverride<A3, int, int, int, int> {
-  const A3Family._()
+  A3Family._()
     : super(
         retry: null,
         name: r'a3Provider',
@@ -203,19 +201,17 @@ abstract class _$A3 extends $Notifier<int> {
 }
 
 @ProviderFor(A4)
-const a4Provider = A4Family._();
+final a4Provider = A4Family._();
 
 final class A4Provider extends $NotifierProvider<A4, int> {
-  const A4Provider._({
-    required A4Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a4Provider',
-         isAutoDispose: false,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A4Provider._({required A4Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a4Provider',
+        isAutoDispose: false,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a4Hash();
@@ -254,7 +250,7 @@ String _$a4Hash() => r'cdd9ad09099881cafe06d7b3095a8b06dbe7d876';
 
 final class A4Family extends $Family
     with $ClassFamilyOverride<A4, int, int, int, int> {
-  const A4Family._()
+  A4Family._()
     : super(
         retry: null,
         name: r'a4Provider',
@@ -292,19 +288,17 @@ abstract class _$A4 extends $Notifier<int> {
 }
 
 @ProviderFor(A5)
-const a5Provider = A5Family._();
+final a5Provider = A5Family._();
 
 final class A5Provider extends $AsyncNotifierProvider<A5, int> {
-  const A5Provider._({
-    required A5Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a5Provider',
-         isAutoDispose: true,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A5Provider._({required A5Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a5Provider',
+        isAutoDispose: true,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a5Hash();
@@ -335,7 +329,7 @@ String _$a5Hash() => r'c83634c22b6a9149aa8787e45c3b7cd6c88b5958';
 
 final class A5Family extends $Family
     with $ClassFamilyOverride<A5, AsyncValue<int>, int, FutureOr<int>, int> {
-  const A5Family._()
+  A5Family._()
     : super(
         retry: null,
         name: r'a5Provider',
@@ -373,19 +367,17 @@ abstract class _$A5 extends $AsyncNotifier<int> {
 }
 
 @ProviderFor(A6)
-const a6Provider = A6Family._();
+final a6Provider = A6Family._();
 
 final class A6Provider extends $AsyncNotifierProvider<A6, int> {
-  const A6Provider._({
-    required A6Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a6Provider',
-         isAutoDispose: false,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A6Provider._({required A6Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a6Provider',
+        isAutoDispose: false,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a6Hash();
@@ -416,7 +408,7 @@ String _$a6Hash() => r'fe641c72cacf3dd119eb77a34fe8fc71c5c30139';
 
 final class A6Family extends $Family
     with $ClassFamilyOverride<A6, AsyncValue<int>, int, FutureOr<int>, int> {
-  const A6Family._()
+  A6Family._()
     : super(
         retry: null,
         name: r'a6Provider',
@@ -454,19 +446,17 @@ abstract class _$A6 extends $AsyncNotifier<int> {
 }
 
 @ProviderFor(A7)
-const a7Provider = A7Family._();
+final a7Provider = A7Family._();
 
 final class A7Provider extends $StreamNotifierProvider<A7, int> {
-  const A7Provider._({
-    required A7Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a7Provider',
-         isAutoDispose: true,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A7Provider._({required A7Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a7Provider',
+        isAutoDispose: true,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a7Hash();
@@ -497,7 +487,7 @@ String _$a7Hash() => r'd3d9ab5090e21987d65522f14ebb70d0058fc56a';
 
 final class A7Family extends $Family
     with $ClassFamilyOverride<A7, AsyncValue<int>, int, Stream<int>, int> {
-  const A7Family._()
+  A7Family._()
     : super(
         retry: null,
         name: r'a7Provider',
@@ -535,19 +525,17 @@ abstract class _$A7 extends $StreamNotifier<int> {
 }
 
 @ProviderFor(A8)
-const a8Provider = A8Family._();
+final a8Provider = A8Family._();
 
 final class A8Provider extends $StreamNotifierProvider<A8, int> {
-  const A8Provider._({
-    required A8Family super.from,
-    required int super.argument,
-  }) : super(
-         retry: null,
-         name: r'a8Provider',
-         isAutoDispose: false,
-         dependencies: null,
-         $allTransitiveDependencies: null,
-       );
+  A8Provider._({required A8Family super.from, required int super.argument})
+    : super(
+        retry: null,
+        name: r'a8Provider',
+        isAutoDispose: false,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$a8Hash();
@@ -578,7 +566,7 @@ String _$a8Hash() => r'54f4a841a283161bed3d444dcee53bf367958678';
 
 final class A8Family extends $Family
     with $ClassFamilyOverride<A8, AsyncValue<int>, int, Stream<int>, int> {
-  const A8Family._()
+  A8Family._()
     : super(
         retry: null,
         name: r'a8Provider',
@@ -616,10 +604,10 @@ abstract class _$A8 extends $StreamNotifier<int> {
 }
 
 @ProviderFor(B)
-const bProvider = BProvider._();
+final bProvider = BProvider._();
 
 final class BProvider extends $NotifierProvider<B, int> {
-  const BProvider._()
+  BProvider._()
     : super(
         from: null,
         argument: null,
@@ -668,10 +656,10 @@ abstract class _$B extends $Notifier<int> {
 }
 
 @ProviderFor(B2)
-const b2Provider = B2Provider._();
+final b2Provider = B2Provider._();
 
 final class B2Provider extends $NotifierProvider<B2, int> {
-  const B2Provider._()
+  B2Provider._()
     : super(
         from: null,
         argument: null,