+++ packages/riverpod_generator/CHANGELOG.md
@@ -1,6 +1,8 @@
-## Unreleased patch
+## Unreleased major
 
+- Generated providers are no-longer constant
 - Fix a bug when using `Class.staticVariable` as parameter to `@Riverpod(retry: ...)`
+- Fix Notifiers returning a `Future<void>`
 
 ## 3.0.3 - 2025-10-09